Tighten exception hierarchy in engine and utilities, fix schema flag types
- Replace ValueError with ConfigError in pipeline_runner.py for invalid
  only parameter, row_group_size format check, and negative promotion
  counts (3 sites)
- Replace KeyError with DimensionError in dimensions_runner.py for
  unknown dimension dependency
- Replace ValueError with ConfigError in trend_presets.py for unknown
  preset names, bad seasonality, and invalid first_year_pct (6 sites)
- Remove dead returns_cfg variable in config.py apply_cross_section_rules
- Remove unused REAL() function from static_schemas.py
- Change 15 boolean flag columns from INT/TINYINT to BIT in static
  schemas: IsOrderDelayed, IsFragile, IsHazmat, EligibleStore,
  EligibleOnline, EligibleMarketplace, EligibleB2B, IsWeekend,
  IsBusinessDay, IsYearStart, IsYearEnd, IsQuarterStart, IsQuarterEnd,
  IsMonthStart, IsMonthEnd
- Align dates generators (calendar.py, fiscal.py, weekly_fiscal.py) to
  produce BIT-schema columns as int8 instead of int32 (15 sites)
- Update tests to expect ConfigError instead of ValueError

+++ src/dimensions/dates/fiscal.py
@@ -56,10 +56,10 @@ def add_fiscal_columns(
     ).dt.normalize()
     df["FiscalQuarterEndDate"] = (df["FiscalQuarterStartDate"] + pd.DateOffset(months=3) - pd.Timedelta(days=1)).dt.normalize()
 
-    df["IsFiscalYearStart"] = (df["Date"] == df["FiscalYearStartDate"]).astype(np.int32)
-    df["IsFiscalYearEnd"] = (df["Date"] == df["FiscalYearEndDate"]).astype(np.int32)
-    df["IsFiscalQuarterStart"] = (df["Date"] == df["FiscalQuarterStartDate"]).astype(np.int32)
-    df["IsFiscalQuarterEnd"] = (df["Date"] == df["FiscalQuarterEndDate"]).astype(np.int32)
+    df["IsFiscalYearStart"] = (df["Date"] == df["FiscalYearStartDate"]).astype(np.int8)
+    df["IsFiscalYearEnd"] = (df["Date"] == df["FiscalYearEndDate"]).astype(np.int8)
+    df["IsFiscalQuarterStart"] = (df["Date"] == df["FiscalQuarterStartDate"]).astype(np.int8)
+    df["IsFiscalQuarterEnd"] = (df["Date"] == df["FiscalQuarterEndDate"]).astype(np.int8)
 
     df["FiscalYear"] = fiscal_year_end.astype(np.int32)
     df["FiscalYearLabel"] = "FY " + df["FiscalYear"].astype(str)
